refactor: split-off Uprobe
* change core to be a folder
* move `Uprobe` to its own module within core
